fixed internalinseconds variable to set like double

+++ Models/RouteInfo.cs
@@ -6,11 +6,11 @@ public class RouteInfo
     public DateTime TravelStartTime { get; }
     public double SpeedKmHr { get; }
     public double RouteDistanceKm { get; }
-    public long NotificationIntervalInSeconds { get; }
+    public double NotificationIntervalInSeconds { get; }
     public Action<PositionNotification> CallBack { get;  }
 
     public RouteInfo(int routeId, PositionTrackerLatLong destination, DateTime travelStartTime, 
-        double speedKmHr, double routeDistanceKm, long notificationIntervalInSeconds, 
+        double speedKmHr, double routeDistanceKm, double notificationIntervalInSeconds, 
         Action<PositionNotification> callBack)
     {
         RouteId = routeId;
